deduplicate createDeployWallet & setupSdk
