+++ universal-login-sdk/test/helpers/createDeployedWallet.ts
@@ -1,11 +1,18 @@
 import {Wallet, utils} from 'ethers';
-import {TEST_GAS_PRICE, ETHER_NATIVE_TOKEN} from '@universal-login/commons';
-import UniversalLoginSDK from '../../src';
+import {WalletService} from '../../src/core/services/WalletService';
+import UniversalLoginSDK, {DeployedWallet} from '../../src';
+import {ETHER_NATIVE_TOKEN, TEST_GAS_PRICE} from '@universal-login/commons';
 
-export const createDeployedWallet = async (ensName: string, sdk: UniversalLoginSDK, deployer: Wallet) => {
+export const createdDeployedWallet = async (name: string, sdk: UniversalLoginSDK, wallet: Wallet): Promise<DeployedWallet> => {
   const {contractAddress, waitForBalance, deploy} = await sdk.createFutureWallet();
-  await deployer.sendTransaction({to: contractAddress, value: utils.parseEther('2.0')});
+  await wallet.sendTransaction({to: contractAddress, value: utils.parseEther('2.0')});
   await waitForBalance();
-  const {waitToBeSuccess} = await deploy(ensName, TEST_GAS_PRICE, ETHER_NATIVE_TOKEN.address);
+  const {waitToBeSuccess} = await deploy(name, TEST_GAS_PRICE, ETHER_NATIVE_TOKEN.address);
   return waitToBeSuccess();
+};
+
+export const createAndSetWallet = async (name: string, walletService: WalletService, wallet: Wallet, sdk: UniversalLoginSDK) => {
+  const {privateKey, contractAddress} = await createdDeployedWallet(name, sdk, wallet);
+  walletService.setWallet({privateKey, contractAddress, name});
+  return {contractAddress, privateKey, walletService};
 };
