--- universal-login-sdk/test/helpers/createWallet.ts
@@ -1,18 +1,0 @@
-import {Wallet, utils} from 'ethers';
-import {WalletService} from '../../src/core/services/WalletService';
-import UniversalLoginSDK, {DeployedWallet} from '../../src';
-import {ETHER_NATIVE_TOKEN, TEST_GAS_PRICE} from '@universal-login/commons';
-
-export const createWallet = async (name: string, sdk: UniversalLoginSDK, wallet: Wallet): Promise<DeployedWallet> => {
-  const {contractAddress, waitForBalance, deploy} = await sdk.createFutureWallet();
-  await wallet.sendTransaction({to: contractAddress, value: utils.parseEther('2.0')});
-  await waitForBalance();
-  const {waitToBeSuccess} = await deploy(name, TEST_GAS_PRICE, ETHER_NATIVE_TOKEN.address);
-  return waitToBeSuccess();
-};
-
-export const createAndSetWallet = async (name: string, walletService: WalletService, wallet: Wallet, sdk: UniversalLoginSDK) => {
-  const {privateKey, contractAddress} = await createWallet(name, sdk, wallet);
-  walletService.setWallet({privateKey, contractAddress, name});
-  return {contractAddress, privateKey, walletService};
-};
